requirements.txt
streamlit_book
pandas
streamlit
plotly
pymongo
wordcloud
Pillow
st-annotated-text